requirements.txt
# Core dependencies
anthropic>=0.7.0
pyyaml>=6.0

# Database


# PPT Generation (NanoBanana)
google-genai>=0.2.0

# Optional: For advanced features
# requests>=2.28.0  # If you add web fetching
# pandas>=1.5.0     # For data analysis
